several minor changes and refactoring to adhere to common guidelines
